add reviewers controller

+++ GlattReviews.Domain/Entities/Contact.cs
@@ -9,6 +9,6 @@ namespace GlattReviews.Domain.Entities
     public class Contact
     {
         public string Email { get; set; }
-        public int PhoneNumber { get; set; }
+        public long PhoneNumber { get; set; }
     }
 }
